requirements.txt
telethon
tkcalendar
Pillow